+++ didman/didman_test.go
@@ -929,17 +929,6 @@ func TestKeyedMutex_Lock(t *testing.T) {
 	})
 }
 
-func TestGenerateIDForService(t *testing.T) {
-	u, _ := url.Parse("https://api.example.com/v1")
-	expectedID := ssi.MustParseURI(fmt.Sprintf("%s#D4eNCVjdtGaeHYMdjsdYHpTQmiwXtQKJmE9QSwwsKKzy", vdr.TestDIDA.String()))
-
-	id := generateIDForService(testDIDA, did.Service{
-		Type:            "type",
-		ServiceEndpoint: u.String(),
-	})
-	assert.Equal(t, expectedID, id)
-}
-
 func TestReferencedService(t *testing.T) {
 	trueDID := did.MustParseDID("did:nuts:123")
 	falseDID := did.MustParseDID("did:nuts:abc")
